The main funtion in order to retrieve the data from the iffi

Generate a shapefile from a given spatial data-object and attributes as produced by a query on the iffi database

make_shapefile(
  database_dir = NULL,
  attribute_database_name = "tbl_frane",
  dictionary_database_name = "diz_frane",
  shapefile = NULL,
  attri = NULL,
  joins = NULL
)

Arguments

database_dir
Path to the directory of the databases

attribute_database_name
Name without extension of the attributes database

dictionary_database_name
Name without extension of the dictionary database

shapefile
A shape that contains the iffi_kodex in a column called PIFF_ID

attri
A vector of attributes to query directly from the attributes database

join
A named list of the tables that need to be joined in order to get the desciptions from the dictionary

Value

An sf-oject

Examples